Move sizeCalculator middleware tests to the correct place
