+++ test/unit/middleware/sizeCalculator.js
