+++ test/unit/run.js
@@ -36,7 +36,6 @@ var tests = [
   require('./helper/geojsonify'),
   require('./helper/logging'),
   require('./helper/type_mapping'),
-  require('./helper/sizeCalculator'),
   require('./helper/stackTraceLine'),
   require('./middleware/access_log'),
   require('./middleware/accuracy'),
@@ -52,6 +51,7 @@ var tests = [
   require('./middleware/parseBBox'),
   require('./middleware/sendJSON'),
   require('./middleware/normalizeParentIds'),
+  require('./middleware/sizeCalculator'),
   require('./middleware/sortResponseData'),
   require('./middleware/trimByGranularity'),
   require('./middleware/trimByGranularityStructured'),
